TIP-874: move AttributeTypeInterface
