+++ src/Akeneo/Pim/Structure/Component/AttributeType/AbstractAttributeType.php
@@ -2,7 +2,7 @@
 
 namespace Akeneo\Pim\Structure\Component\AttributeType;
 
-use Pim\Component\Catalog\AttributeTypeInterface;
+use Akeneo\Pim\Structure\Component\AttributeTypeInterface;
 use Pim\Component\Catalog\AttributeTypes;
 
 /**
