+++ src/Akeneo/Pim/Structure/Component/AttributeTypeRegistry.php
@@ -2,8 +2,6 @@
 
 namespace Akeneo\Pim\Structure\Component;
 
-use Pim\Component\Catalog\AttributeTypeInterface;
-
 /**
  * The attribute type registry
  *
